requirements.txt
# API clients
openai>=1.0.0
google-api-python-client>=2.0.0
youtube-transcript-api>=0.6.0
requests>=2.31.0

# Embeddings + Vector DB
sentence-transformers>=2.2.0
chromadb>=0.4.0

# API server
fastapi>=0.110.0
uvicorn>=0.29.0
fastapi[standard]

# Utilities
python-dotenv>=1.0.0